+++ lib/clickable_widget.dart
@@ -5,3 +5,6 @@ export 'src/clickable_image.dart';
 export 'src/clickable_sized_box.dart';
 export 'src/clickable_card.dart';
 export 'src/clickable_grid_tile.dart';
+export 'src/clickable_column.dart';
+export 'src/clickable_row.dart';
+export 'src/clickable_text.dart';