+++ lib/src/clickable_column.dart
@@ -1,0 +1,147 @@
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+
+/// A clickable column wrapped inside GestureDetector
+class ClickableColumn extends StatelessWidget {
+  /// A pointer that might cause a tap with a primary button has contacted the
+  /// screen at a particular location.
+  final GestureTapDownCallback? onTapDown;
+
+  /// A pointer that will trigger a tap with a primary button has stopped
+  /// contacting the screen at a particular location.
+  final GestureTapUpCallback? onTapUp;
+
+  /// A tap with a primary button has occurred.
+  final GestureTapCallback? onTap;
+
+  /// The pointer that previously triggered [onTapDown] will not end up causing
+  /// a tap.
+  final GestureTapCancelCallback? onTapCancel;
+
+  /// A pointer that might cause a double tap has contacted the screen at a
+  /// particular location.
+  final GestureTapDownCallback? onDoubleTapDown;
+
+  /// The user has tapped the screen with a primary button at the same location
+  /// twice in quick succession.
+  final GestureTapCallback? onDoubleTap;
+
+  /// The pointer that previously triggered [onDoubleTapDown] will not end up
+  /// causing a double tap.
+  final GestureTapCancelCallback? onDoubleTapCancel;
+
+  /// The pointer has contacted the screen with a primary button, which might
+  /// be the start of a long-press.
+
+  final GestureLongPressDownCallback? onLongPressDown;
+
+  /// A pointer that previously triggered [onLongPressDown] will not end up
+  /// causing a long-press.
+
+  final GestureLongPressCancelCallback? onLongPressCancel;
+
+  /// Called when a long press gesture with a primary button has been recognized.
+
+  final GestureLongPressCallback? onLongPress;
+
+  /// Called when a long press gesture with a primary button has been recognized.
+
+  final GestureLongPressStartCallback? onLongPressStart;
+
+  /// A pointer has been drag-moved after a long-press with a primary button.
+  final GestureLongPressMoveUpdateCallback? onLongPressMoveUpdate;
+
+  /// A pointer that has triggered a long-press with a primary button has
+  /// stopped contacting the screen.
+  final GestureLongPressUpCallback? onLongPressUp;
+
+  /// A pointer that has triggered a long-press with a primary button has
+  /// stopped contacting the screen.
+  final GestureLongPressEndCallback? onLongPressEnd;
+
+  // The widgets to display inside the Row
+  final List<Widget> children;
+
+  /// How the children should be placed along the main axis.
+  final MainAxisAlignment mainAxisAlignment;
+
+  /// How much space should be occupied in the main axis.
+  final MainAxisSize mainAxisSize;
+
+  /// How the children should be placed along the cross axis.
+  final CrossAxisAlignment crossAxisAlignment;
+
+  /// Determines the order to lay children out horizontally and how to interpret
+  /// `start` and `end` in the horizontal direction.
+  final TextDirection? textDirection;
+
+  /// Determines the order to lay children out vertically and how to interpret
+  /// `start` and `end` in the vertical direction.
+  final VerticalDirection verticalDirection;
+
+  /// If aligning items according to their baseline, which baseline to use.
+  final TextBaseline? textBaseline;
+
+  /// {@macro flutter.material.Material.clipBehavior}
+  /// Defaults to [Clip.none].
+  final Clip clipBehavior;
+
+  const ClickableColumn({
+    super.key,
+    required this.children,
+    this.textDirection,
+    this.textBaseline,
+    this.clipBehavior = Clip.none,
+    this.mainAxisSize = MainAxisSize.max,
+    this.verticalDirection = VerticalDirection.down,
+    this.crossAxisAlignment = CrossAxisAlignment.center,
+    this.mainAxisAlignment = MainAxisAlignment.start,
+
+    // gesture handler
+    this.onTap,
+    this.onTapUp,
+    this.onTapDown,
+    this.onTapCancel,
+    this.onDoubleTap,
+    this.onDoubleTapDown,
+    this.onDoubleTapCancel,
+    this.onLongPress,
+    this.onLongPressDown,
+    this.onLongPressCancel,
+    this.onLongPressStart,
+    this.onLongPressUp,
+    this.onLongPressEnd,
+    this.onLongPressMoveUpdate,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: onTap,
+      onTapDown: onTapDown,
+      onTapUp: onTapUp,
+      onTapCancel: onTapCancel,
+      onDoubleTapDown: onDoubleTapDown,
+      onDoubleTap: onDoubleTap,
+      onDoubleTapCancel: onDoubleTapCancel,
+      onLongPressDown: onLongPressDown,
+      onLongPressCancel: onLongPressCancel,
+      onLongPress: onLongPress,
+      onLongPressStart: onLongPressStart,
+      onLongPressMoveUpdate: onLongPressMoveUpdate,
+      onLongPressUp: onLongPressUp,
+      onLongPressEnd: onLongPressEnd,
+      child: Column(
+        key: key,
+        mainAxisAlignment: mainAxisAlignment,
+        mainAxisSize: mainAxisSize,
+        crossAxisAlignment: crossAxisAlignment,
+        textDirection: textDirection,
+        verticalDirection: verticalDirection,
+        textBaseline: textBaseline,
+        children: children,
+        
+      ),
+    );
+  }
+}