+++ lib/src/clickable_text.dart
@@ -1,0 +1,179 @@
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+
+/// A clickable text wrapped inside GestureDetector
+/// Very useful to make a link inside a bloc of text
+class ClickableText extends StatelessWidget {
+  /// A pointer that might cause a tap with a primary button has contacted the
+  /// screen at a particular location.
+  final GestureTapDownCallback? onTapDown;
+
+  /// A pointer that will trigger a tap with a primary button has stopped
+  /// contacting the screen at a particular location.
+  final GestureTapUpCallback? onTapUp;
+
+  /// A tap with a primary button has occurred.
+  final GestureTapCallback? onTap;
+
+  /// The pointer that previously triggered [onTapDown] will not end up causing
+  /// a tap.
+  final GestureTapCancelCallback? onTapCancel;
+
+  /// A pointer that might cause a double tap has contacted the screen at a
+  /// particular location.
+  final GestureTapDownCallback? onDoubleTapDown;
+
+  /// The user has tapped the screen with a primary button at the same location
+  /// twice in quick succession.
+  final GestureTapCallback? onDoubleTap;
+
+  /// The pointer that previously triggered [onDoubleTapDown] will not end up
+  /// causing a double tap.
+  final GestureTapCancelCallback? onDoubleTapCancel;
+
+  /// The pointer has contacted the screen with a primary button, which might
+  /// be the start of a long-press.
+
+  final GestureLongPressDownCallback? onLongPressDown;
+
+  /// A pointer that previously triggered [onLongPressDown] will not end up
+  /// causing a long-press.
+
+  final GestureLongPressCancelCallback? onLongPressCancel;
+
+  /// Called when a long press gesture with a primary button has been recognized.
+
+  final GestureLongPressCallback? onLongPress;
+
+  /// Called when a long press gesture with a primary button has been recognized.
+
+  final GestureLongPressStartCallback? onLongPressStart;
+
+  /// A pointer has been drag-moved after a long-press with a primary button.
+  final GestureLongPressMoveUpdateCallback? onLongPressMoveUpdate;
+
+  /// A pointer that has triggered a long-press with a primary button has
+  /// stopped contacting the screen.
+  final GestureLongPressUpCallback? onLongPressUp;
+
+  /// A pointer that has triggered a long-press with a primary button has
+  /// stopped contacting the screen.
+  final GestureLongPressEndCallback? onLongPressEnd;
+
+  /// The text to display
+  final String data;
+
+  /// If non-null, the style to use for this text.
+  final TextStyle? style;
+
+  /// {@macro flutter.painting.textPainter.strutStyle}
+  final StrutStyle? strutStyle;
+
+  /// How the text should be aligned horizontally.
+  final TextAlign? textAlign;
+
+  /// The directionality of the text.
+  final TextDirection? textDirection;
+
+  /// Used to select a font when the same Unicode character can
+  /// be rendered differently, depending on the locale.
+  final Locale? locale;
+
+  /// Whether the text should break at soft line breaks.
+  final bool? softWrap;
+
+  /// How visual overflow should be handled.
+  final TextOverflow? overflow;
+
+  /// The number of font pixels for each logical pixel.
+  final double? textScaleFactor;
+
+  /// An optional maximum number of lines for the text to span, wrapping if necessary.
+  /// If the text exceeds the given number of lines, it will be truncated according
+  /// to [overflow].
+  final int? maxLines;
+
+  /// {@template flutter.widgets.Text.semanticsLabel}
+  /// An alternative semantics label for this text.
+  final String? semanticsLabel;
+
+  /// {@macro flutter.painting.textPainter.textWidthBasis}
+  final TextWidthBasis? textWidthBasis;
+
+  /// {@macro dart.ui.textHeightBehavior}
+  final TextHeightBehavior? textHeightBehavior;
+
+  /// The color to use when painting the selection.
+  final Color? selectionColor;
+
+  const ClickableText(
+    this.data, {
+    super.key,
+    this.style,
+    this.strutStyle,
+    this.textAlign,
+    this.textDirection,
+    this.locale,
+    this.softWrap,
+    this.overflow,
+    this.textScaleFactor,
+    this.maxLines,
+    this.semanticsLabel,
+    this.textWidthBasis,
+    this.textHeightBehavior,
+    this.selectionColor,
+
+    // gesture handler
+    this.onTap,
+    this.onTapUp,
+    this.onTapDown,
+    this.onTapCancel,
+    this.onDoubleTap,
+    this.onDoubleTapDown,
+    this.onDoubleTapCancel,
+    this.onLongPress,
+    this.onLongPressDown,
+    this.onLongPressCancel,
+    this.onLongPressStart,
+    this.onLongPressUp,
+    this.onLongPressEnd,
+    this.onLongPressMoveUpdate,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: onTap,
+      onTapDown: onTapDown,
+      onTapUp: onTapUp,
+      onTapCancel: onTapCancel,
+      onDoubleTapDown: onDoubleTapDown,
+      onDoubleTap: onDoubleTap,
+      onDoubleTapCancel: onDoubleTapCancel,
+      onLongPressDown: onLongPressDown,
+      onLongPressCancel: onLongPressCancel,
+      onLongPress: onLongPress,
+      onLongPressStart: onLongPressStart,
+      onLongPressMoveUpdate: onLongPressMoveUpdate,
+      onLongPressUp: onLongPressUp,
+      onLongPressEnd: onLongPressEnd,
+      child: Text(
+        data,
+        key: key,
+        style: style,
+        strutStyle: strutStyle,
+        textAlign: textAlign,
+        textDirection: textDirection,
+        locale: locale,
+        softWrap: softWrap,
+        overflow: overflow,
+        textScaleFactor: textScaleFactor,
+        maxLines: maxLines,
+        semanticsLabel: semanticsLabel,
+        textWidthBasis: textWidthBasis,
+        textHeightBehavior: textHeightBehavior,
+        selectionColor: selectionColor,
+      ),
+    );
+  }
+}